Add a way to check run-time constraints of FSMs
Four types of constraints are possible, and are demonstrated in
`tests/lib.rs` in the `test_check()` function. There is a macro for each
type to allow declaring a given constraint. The four macros are as
follows:

* precondition! - tests that a given predicate on the context of the fsm
                  is true when a state is entered.
* postcondition! - tests that a given predicate on the context of the
                   fsm is true when a state is exited
* invariant! - tests that a given predicate on the context of the fsm is
               always true no matter the current state of the fsm
* transition! - tests that a given predicate on the context of the fsm
                is true when an fsm transitions from one state to the
                next

Once a use declares the constraints, they must run the fsm checker which
runs the local fsm with a given set of messages and validates each
constraint during runtime. This is all demonstrated in the
`test_check()` function.