Add basic Channel and Stage support
Stages can be constructed and stored in a map as &Stage. They can then
be used to get their senders as Box<MsgSender> thus abstracting sending
messages to a stage from how the stage is implemented.

Still need to refactor code to make the checker work with stages and
channels and not just standalone fsms.