Allow FSM states to return output from next!()
Previously, FSMs communicated with the each other and the outside world via channels passed
in as part of the context. Now when an FSM state function calls `next!` it can
optionally return output. This output can be messages meant to be routed to
other FSMs or sent on the wire.

The FsmHandler trait was renamed to FsmTypes and now only contains the
types used to parameterize the FSM module. There is no longer an
`initial_state()` function because of this, so the initial state is now
passed into the FSM constructor. Due to the change in the `next!` macro
so that it returns the next state and output, the initial state is
constructed using the new `state_fn` macro.